[rtd-ms-transaction-filter/develop] Updated InboundTransactionItemProcessorTest and LineAwareMapperTest

+++ api/batch/src/test/java/it/gov/pagopa/rtd/transaction_filter/batch/mapper/LineAwareMapperTest.java
@@ -87,7 +87,7 @@ public class LineAwareMapperTest {
                 .idTrxAcquirer("1111111111")
                 .idTrxIssuer("5555")
                 .correlationId("")
-                .amount(BigDecimal.valueOf(1111))
+                .amount(1111L)
                 .amountCurrency("896")
                 .acquirerId("22222")
                 .merchantId("0000")
